chore: rename Card component
